+++ nullstone.tf
@@ -4,10 +4,10 @@ terraform {
       source = "nullstone-io/ns"
     }
     google = {
-      source  = "hashicorp/google"
+      source = "hashicorp/google"
     }
     google-beta = {
-      source  = "hashicorp/google-beta"
+      source = "hashicorp/google-beta"
     }
   }
 }